oauth2-authenticator/pyproject.toml
[tool.poetry]
name = "oauth2-authenticator"
version = "0.1.0"
description = "arXiv user admin API"
authors = ["arXiv <developer@arxiv.org>"]
license = "BSD-3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.115.11"
arxiv-base = {git = "https://github.com/arXiv/arxiv-base.git", rev = "cdc6fcb271ac79ae44c741c7e664447e86849546"}
pymysql = "^1.1.0"
ruamel-yaml = "^0.18.6"
python-json-logger = "^2.0.7"
python-dateutil = "^2.8.2"
pyjwt = "^2.9.0"
requests = "^2.32.3"
jwcrypto = "^1.5.6"
python-keycloak = "^5.1.1"
httpx = "^0.28.1"
pycountry = "^24.6.1"
captcha = "^0.6.0"
lameenc = "^1.8.1"
arxiv-bizlogic = {git = "https://github.com/arXiv/arxiv-keycloak.git", rev = "master", subdirectory = "bizlogic"}
asgi-correlation-id = "^4.3.4"

[tool.poetry.group.dev.dependencies]
uvicorn = "*"
mypy = "^1.15.0"
pytest = "^8.3.2"
selenium = "^4.27.1"
types-requests = "^2.32.0.20241016"
pytest-mock = "^3.14.0"
dotenv = "^0.9.9"
pytest-asyncio = "^0.26.0"
datamodel-code-generator = "^0.28.5"
openapi-python-client = "^0.24.3"
types-pytz = "^2025.2.0.20250326"
types-python-dateutil = "^2.9.0.20241206"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
